seems to be working

+++ MCT.h
@@ -13,7 +13,7 @@
 class MCT
 {
     public:
-        MCT(float m1,float m2,float a1,float a2)
+        MCT(double m1,double m2,double a1,double a2)
         {
             set_m1(m1);
             set_m2(m2);
@@ -23,14 +23,14 @@ class MCT
 
         ~MCT(){}
 
-        void set_m1(float m1) {this->m1=m1;}
-        void set_m2(float m2) {this->m2=m2;}
-        void set_a1(float a1) {this->a1=a1;}
-        void set_a2(float a2) {this->a2=a2;}
-        void set_m1_predict(float m1_predict){this->m1=m1_predict;}
-        void set_m2_predict(float m2_predict) {this->m2=m2_predict;}
+        void set_m1(double m1) {this->m1=m1;}
+        void set_m2(double m2) {this->m2=m2;}
+        void set_a1(double a1) {this->a1=a1;}
+        void set_a2(double a2) {this->a2=a2;}
+        void set_m1_predict(double m1_predict){this->m1=m1_predict;}
+        void set_m2_predict(double m2_predict) {this->m2=m2_predict;}
         
-        Eigen::MatrixXf M(float q1,float q2)
+        Eigen::MatrixXf M(double q1,double q2)
         {
             Eigen::MatrixXf M(2,2);
             M(0,0)= a1*a1*m1 + (a1*a1 + a2*a2 +2*a1*a2*cos(q2))*m2;
@@ -39,7 +39,7 @@ class MCT
             M(1,1)= a2*a2*m2;
             return M;
         }
-        Eigen::MatrixXf C(float q1,float q2, float q1dot, float q2dot)
+        Eigen::MatrixXf C(double q1,double q2, double q1dot, double q2dot)
         {
             Eigen::MatrixXf C(2,2);
             C(0,0)=-a1*a2*m2*sin(q2)*q2dot;
@@ -48,7 +48,7 @@ class MCT
             C(1,1)= 0;
             return C;
         }
-        Eigen::MatrixXf T(float q1, float q2)
+        Eigen::MatrixXf T(double q1, double q2)
         {
             Eigen::MatrixXf T(2,1);
             T(0,0)=a1*m1*g*cos(q1) +a1*m2*g*cos(q1) +a2*m2*g*cos(q1+q2);
@@ -58,9 +58,9 @@ class MCT
 
     
     private:
-        float m1= 3;
-        float m2= 2;
-        float a1= 2;
-        float a2= 3;
-        float g= 9.8;
+        double m1= 3;
+        double m2= 2;
+        double a1= 2;
+        double a2= 3;
+        double g= 9.8;
 };
